DIG-69: Minification w/obfuscation
So far the generator has been supporting standard minification, which was on by
default. Due to feedback it has been decided to integrate a obfuscation step with
the minification process. However, minification is now off by default. The
initial `package.json` section looks now like this:

```json
  "dizmo": {
    "build": {
      "lint": true,
      "minify": false
    },
  },
```

When the `minify` flag is set to `true` then all HTML, (S)CSS and JS is minified
using state of the art tools. Further, the JS is obfuscated using the popular and
powerful [javascript-obfuscator][1]. The `minify` flag is smart, and can fully be
expanded to really tweak and control the minification/obfuscation process:

```json
  "dizmo": {
    "build": {
      "lint": true,
      "minify": {
        "markups": false,
        "scripts": false,
        "styles": false
      }
    },
  },
```

Above, each minification process for the HTML, (S)CSS and JS can now be
controlled separately. Further, each of the corresponding `markups`, `scripts`
and `styles` minification flags can be even more expanded:

```json
  "dizmo": {
    "build": {
      "lint": true,
      "minify": {
        "markups": {
          "htmlmin": false
        },
        "scripts": {
          "sourcemaps": false,
          "obfuscator": false,
          "uglify": false
        },
        "styles": {
          "sourcemaps": false,
          "sass": false
        }
      }
    },
  }
```

Above, the `markups/htmlmin` flag controls the minification process for the HTML
markup. It's off by default and can be toggled to `true`; further, if desired
the `true` value can actually be replaced with a configuration object, which is
documented in detail in the original [HTML minifier][1] project (or in less
detail at corresponding [Gulp plugin][2] page).

Similarly the `scripts/obfuscator`, `scripts/uglify` and `styles/sass` flags are
also smart and are off by default. They can be toggled on with a `true` value,
or again by a specific object controlling exactly the minification (and/or
obfuscation) process. The corresponding documentation are available at:

* https://github.com/javascript-obfuscator/javascript-obfuscator,
* https://www.npmjs.com/package/gulp-uglify, and
* https://www.npmjs.com/package/gulp-sass.

[1]: https://github.com/kangax/html-minifier
[2]: https://www.npmjs.com/package/gulp-htmlmin