DIG-12: `npm run upload` w/env variables
It's possible to to upload a dizmo `*.dzm` artifact to the dizmo store, by
executing `npm run upload`, which expects the follogin three environment
variables to be set correctly:

* DZM_STORE_HOST = 'https://store-stg.dizmo.com'
* DZM_STORE_USER = 'my-user`
* DZM_STORE_PASS = '*******`

Above the staging server has been used as an example. To upload to the production
store server the host URL is required to be set correspondingly.