DIG-105: npm run upload w/publication
Modified `npm run upload` such that anything that is uploaded, is also published.
In the rare cases, where *only* an upload is desired, but *no* publication the
following can be executed:

```
npm run upload -- --no-publish
```

Or equivalently:
```
npm run -- upload --no-publish
```

Or equivalently:
```
npm run-script -- upload --no-publish
```

In cases where *only* a publication is required, but the upload itself is desired
to be skipped, the following can be executed:

```
npm run upload -- --publish
```

This is actualy semantically slightly unclean, since reading the statement above
one could derive that an upload *and* an (explicit) publication is performend,
which however is *not* the case, since *only* a publication is triggered.

However, since the use case where an upload *with* aN automatic publication
should cover 99% of all use cases, the slight semantic inconistency above has
been ignored, since otherwise we would be required to either introduce a separate
`npm run publish` script or a `--publish-only` flag (for `npm run upload`),
neither of which IMHO is really necessary, given the 1% rarity of a separate
upload and publication step.