DIG-75: avoid `npm install` output
Thanks to DIG-70, it is possible now to automatically invoke `npm install`
*implicitly* through a script (e.g. `npm run make`). Hence, running the
installation process upon a project generation is *not* necessary anymore.

Further, the output of `npm install` can be messy and there is not much, a
developer can do here. But again thanks to DIG-70, it's possible now to suppress
the output of `npm install` and simply show a CLI spinner.

The result is, that (1) a project generation is very fast now, since the
dependencies are not immediately fetched. Plus (2), upon running a script the
`npm install` step is executed (*only* if necessary), while the (ugly) output is
replaced  with a CLI spinner.

Points (1) and (2) result in an initial fast user/develper experience upon
generating a project, and the (only initially slower) second step does not at
least produce an ugly terminal output anymore.