DIG-80: yo dizmo --auto-upgrade (for build system)
Since each dizmo contains it's own build systems (beneath `gulp/`) each has to
updated manually (if desired so) to the most recent build system offered by
dizmoGen.

This arrangement allows to to rapidly iterate and improved dizmoGen, while
letting the developer choose, when (or if) he wants to migrate to the most recent
build system.

Since dizmoGen is based on Yeoman, it offers alread an build-in upgrade mechanism
with a correponding conflict resolution, when an older build system is being
overwritten. To use it, dizmoGen needs to be invoked within a project:

```
$ yo dizmo
```

Any file which is being overwritten needs to be signed-off by the developer: This
helps to avoid accidents. However, for relatively new developers using dizmoGen
deciding what needs to be overwritten, and what not can be, at the very beginning
at least, a problem.

These are the rules, to ensure an upgrade with a minimal friction potential:

* Overwrite anything beneath the `gulp/` folder;
* overwrite the `gulpfile.js` script;
* partially overwrite `package.json` (only `devDependencies`);
* leave the rest as it is (and remove `node_modules` deps).

This rules above should cover a majority of the upgrade use cases; otherwise the
developer has to think for himself! However, even though the rules above are very
simple to follow, a *new* developer does not know them. Hence a `--auto-upgrade`
flag has been introduced. Run within an existing project:

```
$ yo dizmo --auto-upgrade
```

Which will *automatically* follow the upgrade rules from above, and ensure a
simple and quick upgrade to the most recent build system offered by dizmoGen.
Please notice, that nothing except `gulp/`, `gulpfile.js` and `package.json` is
modified, whereas the modification of the latter three elements is *automatic*
(i.e. enforced).

This auto-upgrade mechanims has *not* been implemented for the `type-script` and
`coffee-script` sub-generators yet, which will be handled in separate tickets.