DIG-87: Extend `package.json` template w.r.t `dizmo.settings`
Upon major request, in the `package.json` template `dizmo.settings` has been
extended to look like:

```json
  "dizmo": {
    "settings": {
      "attributes": {
        "settings/usercontrols/allowresize": true
      },
      "bundle-identifier": "<%= bundleId %>",
      "bundle-name": "<%= dizmoName %>",
      "category": "",
      "height": 360,
      "tags": [
         "<%= _.kebabCase(dizmoName) %>"
      ],
      "width": 480
    },
    ..
  }
```

Hence, now we have the following *new* entries:

* `attributes`: where `allowresize` is set by default to `true`;
* `tags`: where it is set by default to e.g. `["my-dizmo"]`;

The first change to is inform developers that a dizmos' attributes can already be
changed via a simple configuration in `package.json`. The latter change was to
encourage them to use `tags` for the dizmoStore (where a non-empty `category` is
also required).