DIG-119: Fixed sourcemaps for Type- and CoffeeScript
When creating projects based on Type- or CoffeeScript, then the corresponding sourcemaps were not generated in a
correct way, which as been fixed now.

For example to create a project based on TypeScript (with GIT enabled):
```
$ yo dizmo --git --type-script
```

Then within the project to build it with sourcemaps, run:
```
my-dizmo.git $ npm run -- make --sourcemaps
```

Or when minification is desired but with sourcemaps support, run:
```
my-dizmo.git $ npm run -- make --minify --no-obfuscate --sourcemaps
```

The `--no-obfuscate` flag is required, since providing only `--minify` implies `--obfuscate`. And sourcemaps with
obfuscation do *not* work, since they are *not* supposed to when obfuscation is on!

Instead of using the control flags via the CLI, they can also be controled via `package.json`, although it is
recommeded to avoid changing the default `package.json` settings w.r.t. to the build flags. If in general a non-
default build behaviour is desired, the flags should be set via e.g. `~/.generator-dizmo/config.json`.